+++ docs/api/changelog.rst
@@ -13,7 +13,14 @@ Fixed
 ~~~~~
 - Incorrect validation error ``data values found at nodata values of
   idomain`` where boundary condition packages with a scalar coordinate, which is
-  not set as dimension. 
+  not set as dimension.
+
+Added
+~~~~~
+- The :func:`imod.mf6.model.mask_all_packages` now also masks the idomain array
+  of the model discretization, and can be used with a mask array without a layer
+  dimension, to mask all layers the same way
+
 
 [0.15.3] - 2024-02-22
 ---------------------
@@ -34,8 +41,7 @@ Added
 - An API package was added. It can be added to both flow and transport models, and its 
   presence allows users to interact with libMF6.dll through its API.
 - Developer environment: Empty python 3.10, 3.11, 3.12 environments where pip
-  install and import imod can be tested. 
-
+  install and import imod can be tested.
 
 [0.15.2] - 2024-02-16
 ---------------------